Add files via upload
A palindrome is a word, phrase, number, or other sequence of characters that reads the same forwards and backwards, ignoring spaces, punctuation, and capitalization. For example, the word "madam" is a palindrome because it reads the same way from left to right and from right to left. 
To us this feature, you just a word or phrase in the INPUT and use the CHECK button to see if your input is a Palindrome✌️